YAY-114 refactor: transfer TabsBlock from entities to widgets
