+++ src/widgets/TabsBlock/index.ts
@@ -1,0 +1,1 @@
+export * from './TabsBlock'
